IGNITE-19212 Moved shared header
