+++ modules/platforms/cpp/ignite/client/detail/node_connection.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include <ignite/client/detail/client_operation.h>
+#include "ignite/protocol/client_operation.h"
 #include <ignite/client/detail/protocol_context.h>
 #include <ignite/client/detail/response_handler.h>
 #include <ignite/client/ignite_client_configuration.h>
