+++ modules/platforms/cpp/ignite/protocol/client_operation.h
